Backtrack

Documentation

bt :: Formula -> Variables -> Assignment

This function takes a formula and the set of variables and applies the Backtracking algorithm to this formula, returning a satisfying assignment