Dpll

Documentation

unitPropogate :: Formula -> Formula -> (Assignment, Formula)

Given a formula, searches for a clause with only one literal and returns and assignment which will remove such a unit clause and the new formula

findPure :: Formula -> [Variable] -> [(Variable, Bool)]

Given a formula and the set of variables, this function finds the pure variables in the formula, and returns their corresponding true assignment

dpll :: t -> StateT Formula Identity Bool

Runs the DPLL algorithm with Unit Propogation, moving from one assignment state to another in each iteration