Gsat

Documentation

runMultiple :: (Eq a, Num a) => Formula -> [Variable] -> a -> Int -> Assignment

Runs the GSAT algorithm for a given formula and set of variables, with the two parameters representing the maximum number of tries and flips allowed