Satsolve

Documentation

type Variable = String

data DoubtFul a

Constructors

Mtrue

Mfalse

Assign [a]

data BVariable a

Constructors

Same a

Negate a

type Literal = BVariable Variable

A literal can either be a variable or the negation of a variable

type Clause = [Literal]

A clause is a set of literals

type Formula = DoubtFul Clause

A formula can either be True or False or a set of unsatisfied clauses

type Variables = [Variable]

type Assignment = [(Variable, Bool)]

An Assignment is a list of variables and their boolean values

same :: Formula -> Formula -> Bool

Takes two formulas as input and checks if both are True or both are False

assignValues :: Formula -> Assignment -> (Formula, Assignment)

This function takes a Formula and an Assignment and returns the resultant Formula from applying the Assignment to the initial one

getUnsat :: Formula -> Assignment -> Int

This function takes a Formula and Assignment and returns the number of clauses not satisfied by this assignment

getSat :: Formula -> Assignment -> Int

This function takes a Formula and Assignment and returns the number of clauses satisfied by this assignment

mor :: Formula -> Formula -> Bool

Corresponding to the contructors in Doubtful data-type, this return the OR of the two formulas given

stateFormula :: Formula -> State Assignment Formula

Given a formula, it creates the State monad corresponding to the formula